Mejora las validaciones de formulario. Refactoriza y fixea errores pequeños.
